+++ scripts/nats/manage_streams.py
