+++ scripts/nats/memory_monitor.py
