+++ src/huntsman/pocs/nats/streams.py
@@ -1,27 +1,42 @@
-import argparse
-import asyncio
-from dataclasses import dataclass
+from dataclasses import dataclass, asdict
+from typing import Optional, List, Tuple
 
-import nats
-from nats.js import JetStreamContext
-
-from huntsman.pocs.nats.utils import list_streams
+from nats.js import JetStreamContext, api
 
 
 @dataclass
-class StreamConfig():
-    """Stream Configuration"""
-    memory_bytes: int = 1_000_000_000
-    disk_bytes: int = 10_000_000_000
+class MemoryStreamConfig(api.StreamConfig):
+    """Stream Configuration for memory stream. Gives sensible defaults for the official StreamConfig object"""
+    max_bytes: int = 1_000_000_000
+    retention: str = "workqueue",
+    storage: str = "memory",
+    max_age: Optional[float] = 60,
+    max_msgs: Optional[int] = 10000,
+    discard: str = "new",
+    no_ack: bool = False,
+    duplicate_window: float = 60,
 
 
-async def start_streams(js: JetStreamContext, num_streams: int, cfg: StreamConfig = StreamConfig()) -> None:
+@dataclass
+class DiskStreamConfig(api.StreamConfig):
+    """Stream Configuration for disk stream. Gives sensible defaults for the official StreamConfig object"""
+    max_bytes: int = 10_000_000_000
+    retention: str = "interest",
+    storage: str = "file",
+    max_age: Optional[float] = 120,
+    max_msgs: Optional[int] = 10000,
+    discard: str = "old",
+    no_ack: bool = False,
+
+
+async def start_streams(js: JetStreamContext, num_streams: int, disk_cfg: DiskStreamConfig = DiskStreamConfig(), mem_cfg: MemoryStreamConfig = MemoryStreamConfig()) -> None:
     """Sets up all data streams.
 
     Args:
         js: The JetStreamContext object
-        num_streams: The number os streams (memory and disk) to create
-        cfg: The StreamConfig object used for stream configuration
+        num_streams: The number os streams(memory and disk) to create
+        disk_cfg: The DiskStreamConfig object used for the disk stream configuration
+        mem_cfg: The MemoryStreamConfig object used for the memory stream configuration
     """
     # Connect to NATS
     print(f"Creating {num_streams} memory streams and {num_streams} disk streams...")
@@ -30,18 +45,9 @@ async def start_streams(js: JetStreamContext, num_streams: int, cfg: StreamConfi
     for i in range(num_streams):
         try:
             stream_no = i+1
-            await js.add_stream(
-                name=f"CAMERA_MEMORY_{stream_no}",
-                subjects=[f"camera.memory.{stream_no}.>"],
-                retention="workqueue",
-                storage="memory",
-                max_age=60,
-                max_msgs=10000,
-                max_bytes=cfg.memory_bytes,
-                discard="new",
-                no_ack=False,
-                duplicate_window=60,
-            )
+            mem_cfg.name = f"CAMERA_MEMORY_{stream_no}"
+            mem_cfg.subjects = [f"camera.memory.{stream_no}.>"]
+            await js.add_stream(**asdict(mem_cfg))
             print(f"Created CAMERA_MEMORY_{stream_no} stream")
         except Exception as e:
             print(f"Error creating CAMERA_MEMORY_{stream_no} stream: {e}")
@@ -50,17 +56,9 @@ async def start_streams(js: JetStreamContext, num_streams: int, cfg: StreamConfi
     for i in range(num_streams):
         try:
             stream_no = i+1
-            await js.add_stream(
-                name=f"CAMERA_DISK_{stream_no}",
-                subjects=[f"camera.archive.{stream_no}.>"],
-                retention="interest",
-                storage="file",
-                max_age=120,
-                max_msgs=10000,
-                max_bytes=cfg.disk_bytes,
-                discard="old",
-                no_ack=False,
-            )
+            disk_cfg.name = f"CAMERA_DISK_{stream_no}",
+            disk_cfg.subjects = [f"camera.archive.{stream_no}.>"],
+            await js.add_stream(**asdict(disk_cfg))
             print(f"Created CAMERA_DISK_{stream_no} stream")
         except Exception as e:
             print(f"Error creating CAMERA_DISK_{stream_no} stream: {e}")
@@ -99,64 +97,25 @@ async def delete_streams(js: JetStreamContext) -> None:
     print("Stream deletion complete")
 
 
-async def start(nats_server: str, num_streams: int, cfg: StreamConfig = StreamConfig()) -> None:
-    """Connects to the nats server, creates the Jetstream context and Begins
-    all streams
+async def list_streams(js: JetStreamContext, mem_pattern: str = "CAMERA_MEMORY_", disk_pattern: str = "CAMERA_DISK_") -> Tuple[List[str], List[str]]:
+    """List all streams matching our naming pattern.
 
     Args:
-        nats_server: The hostname of the nats server to start the streams
-        num_streams: The number os streams (memory and disk) to create
-        cfg: The StreamConfig object used for stream configuration
+
     """
-    # Connect to NATS
-    print(f"Connecting to NATS server at {nats_server}")
-    nc = await nats.connect(servers=[nats_server])
-    js = nc.jetstream()
-    try:
-        await start_streams(js, num_streams, cfg)
-    finally:  # Always close connection
-        await nc.close()
+    streams = await js.streams_info()
+    memory_streams = []
+    disk_streams = []
 
+    for stream in streams:
+        name = stream.config.name
+        if name.startswith(mem_pattern):
+            memory_streams.append(name)
+        elif name.startswith(disk_pattern):
+            disk_streams.append(name)
 
-async def delete(nats_server: str):
-    """Deletes all streams connected to a nats server:
+    # Sort streams by their index to match them correctly
+    memory_streams.sort(key=lambda x: int(x.split("_")[-1]))
+    disk_streams.sort(key=lambda x: int(x.split("_")[-1]))
 
-    Args:
-        nats_server: The hostname of the nats server to start the streams
-    """
-    print("Deleting streams...")
-    print(f"Connecting to NATS server at {nats_server}")
-    nc = await nats.connect(servers=[nats_server])
-    js = nc.jetstream()
-    try:
-        await delete_streams(js)
-    finally:
-        await nc.close()
-
-
-async def _main():
-    if args.delete:
-        asyncio.run(delete(nats_server=args.nats_server))
-    else:
-        cfg = StreamConfig(memory_bytes=args.memory_bytes, disk_bytes=args.disk_bytes)
-        asyncio.run(start(nats_server=args.nats_server,
-                          num_streams=args.num_streams, cfg=cfg))
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description="Setup and run a NATS Jetstream streams. Or tear down existing streams.")
-    parser.add_argument("--delete", action="store_true", default=False,
-                        help="Whether to delete streams previously created, instead of creating new streams.")
-    parser.add_argument("-s", "--nats-server", default="nats://localhost:4222",
-                        help="The nats server host and port.")
-    parser.add_argument("-n", "--num-streams", type=int, default=10,
-                        help="The number of streams to create.")
-
-    cfg = StreamConfig()
-    parser.add_argument("-m", "--memory-bytes", default=cfg.memory_bytes, type=int,
-                        help="The size of the memory stream in bytes")
-    parser.add_argument("-d", "--disk-bytes", default=cfg.disk_bytes, type=int,
-                        help="The size of the disk stream in bytes")
-
-    args = parser.parse_args()
-    asyncio.run(_main(args))
+    return memory_streams, disk_streams
